Weather/Rasp: Store all render settings in same profile section Load / store weather render settings together in the map profile section rather than being spread out over profile keys and UI state.
