+++ src/MapSettings.cpp
@@ -51,6 +51,7 @@ MapSettings::SetDefaults() noexcept
   online_traffic_map_mode = DisplayOnlineTrafficMapMode::SYMBOL;
   show_95_percent_rule_helpers = false;
   rasp_layer_opacity = 70;
+  rasp_contour_density = ContourDensity::OFF;
 
   trail.SetDefaults();
   item_list.SetDefaults();
